requirements.txt
openai>=1.0.0
qq-botpy>=1.2.0
